pharma_knowledge_distillation/requirements.txt
# Core ML and AI Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
jieba>=0.42.1

# Web Interface
gradio>=3.35.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API and Networking
requests>=2.31.0
httpx>=0.24.0
ollama>=0.1.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Logging and Monitoring
loguru>=0.7.0
wandb>=0.15.0
tensorboard>=2.13.0

# Data Validation
pydantic>=2.0.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: GPU Support
# cuda-toolkit  # Install separately based on your CUDA version

# Optional: Advanced Features
optuna>=3.3.0  # Hyperparameter optimization
ray[tune]>=2.6.0  # Distributed training
datasets>=2.14.0  # Hugging Face datasets

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Production Deployment
fastapi>=0.100.0
uvicorn>=0.22.0
gunicorn>=21.0.0